Update score type to float and add Python score parser utility

+++ models/score.go
@@ -5,5 +5,5 @@ import "gorm.io/gorm"
 type Score struct {
 	gorm.Model
 	GitRepo string `json:"git_repo" example:"user_name/repo_name" validate:"required"`
-	Score int `json:"score" example:"100" validate:"required"`
+	Score float64 `json:"score" example:"100" validate:"required"`
 }
